# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZEROHOUR REALTOR PULSE API",
    "version": "0.1.0",
    "description": "Phase 1 local MVP API contract for ZEROHOUR REALTOR PULSE. This contract supports frontend integration, simulated creator onboarding, async provisioning, mock ZIP signal ingestion, cached dashboards, signal actions, billing handoff, and basic admin operations."
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local development API"
    }
  ],
...
  "paths": {
    "/auth/signup": {
      "post": {
        "tags": ["auth"],
        "summary": "Create account and enqueue async provisioning",
        "operationId": "signup",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              },
              "example": {
                "email": "agent@example.com",
                "password": "correct-horse-123",
                "full_name": "Avery Agent",
                "zip_codes": ["90210", "10001"],
                "creator_code": "creator_a"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Signup accepted and provisioning queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignupResponse"
                }
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Login with email and password",
        "operationId": "login",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "example": {
                "email": "agent@example.com",
                "password": "correct-horse-123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JWT issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/admin/ingestion/{zip_code}/enqueue": {
      "post": {
        "tags": ["admin"],
        "summary": "Enqueue ZIP ingestion",
        "operationId": "enqueueIngestion",
        "parameters": [
          {
            "name": "zip_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "mock"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ingestion queued"
          }
        }
      }
    },
    "/admin/kill-switches/{key}": {
      "put": {
        "tags": ["admin"],
        "summary": "Set operational kill switch",
        "operationId": "setKillSwitch",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "enabled",
            "in": "query",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "reason",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Kill switch updated"
          }
        }
      }
    },
    "/admin/feature-flags/{key}": {
      "put": {
        "tags": ["admin"],
        "summary": "Set feature flag",
        "operationId": "setFeatureFlag",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "enabled",
            "in": "query",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "description",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Feature flag updated"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "responses": {
...
      "Unauthorized": {
        "description": "Unauthorized"
      },
...
      "Conflict": {
        "description": "Conflict"
      }
    },
    "schemas": {
      "SignupRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 10
          },
          "full_name": {
            "type": ["string", "null"]
          },
          "zip_codes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "maxItems": 20
          },
          "creator_code": {
            "type": ["string", "null"]
          }
        }
      },
      "SignupResponse": {
        "type": "object",
        "required": ["access_token", "token_type", "user_id", "tenant_id", "provisioning_job_id", "onboarding_state"],
        "properties": {
          "access_token": {
            "type": "string"
          },
          "token_type": {
            "type": "string",
            "default": "bearer"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid"
          },
          "provisioning_job_id": {
            "type": "string",
            "format": "uuid"
          },
          "onboarding_state": {
            "type": "string"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "required": ["access_token", "token_type"],
        "properties": {
          "access_token": {
            "type": "string"
          },
          "token_type": {
            "type": "string",
            "default": "bearer"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}